Correct name parsing
https://hyperrealm.github.io/libconfig/libconfig_manual.html#Configuration-Files
says: `All names are case-sensitive. They may consist only of
alphanumeric characters, dashes (‘-’), underscores (‘_’), and asterisks
(‘*’), and must begin with a letter or asterisk. No other characters are
allowed.`

Adjust the key parser to match that specification. This also requires
one of the test to be modified as it was testing with an invalid config
fragment, that actually gets rejected by libconfig.